feat(showcase): large corpus, over_domain_bench, showcase script
config/obs/corpus_large.yaml — 1M ticks, 37 texts, 9 subsystems
  (~60K injections, ~900K–1M tokens)

axiom-bench/benches/over_domain_bench.rs — new V7 benchmark:
  cold / warm-100 / loaded-1000 ticks, throughput (1000 ticks),
  inject cold vs loaded; measures ContextRecognizer V7 + NeuralAdvisor
  + FatigueStore + TransitionMatrix overhead in steady state

hot_path_regression.rs — update comment for V7 pipeline,
  sample_size 1000 → 500

stress_bench.rs — fix i16 overflow in make_positions/make_pos_tuples
  (% 60000 exceeds i16 max; use i32 intermediate)

scripts/showcase.sh — orchestration: build → OBS 1M ticks → benches
  → showcase/SHOWCASE.md; flags --obs-only / --bench-only

justfile — add showcase / showcase-bench / showcase-obs / bench recipes

+++ crates/axiom-bench/benches/stress_bench.rs
@@ -16,10 +16,10 @@ use std::time::Duration;
 fn make_positions(n: usize) -> Vec<[i16; 3]> {
     (0..n)
         .map(|i| {
-            let x = ((i.wrapping_mul(37)) % 60000) as i16 - 30000;
-            let y = ((i.wrapping_mul(53)) % 60000) as i16 - 30000;
-            let z = ((i.wrapping_mul(71)) % 10000) as i16 - 5000;
-            [x, y, z]
+            let x = ((i.wrapping_mul(37)) % 60000) as i32 - 30000;
+            let y = ((i.wrapping_mul(53)) % 60000) as i32 - 30000;
+            let z = ((i.wrapping_mul(71)) % 10000) as i32 - 5000;
+            [x as i16, y as i16, z as i16]
         })
         .collect()
 }
@@ -31,10 +31,10 @@ fn make_masses(n: usize) -> Vec<u16> {
 fn make_pos_tuples(n: usize) -> Vec<(i16, i16, i16)> {
     (0..n)
         .map(|i| {
-            let x = ((i.wrapping_mul(37)) % 60000) as i16 - 30000;
-            let y = ((i.wrapping_mul(53)) % 60000) as i16 - 30000;
-            let z = ((i.wrapping_mul(71)) % 10000) as i16 - 5000;
-            (x, y, z)
+            let x = ((i.wrapping_mul(37)) % 60000) as i32 - 30000;
+            let y = ((i.wrapping_mul(53)) % 60000) as i32 - 30000;
+            let z = ((i.wrapping_mul(71)) % 10000) as i32 - 5000;
+            (x as i16, y as i16, z as i16)
         })
         .collect()
 }
